Move types next to components
